+++ src/gym_longicontrol/envs/__init__.py
@@ -1,3 +1,10 @@
 from .longicontrol import DeterministicTrack, LongiControlEnv, StochasticTrack
+from .multi_objective import MODeterministicTrack, MOStochasticTrack
 
-__all__ = ["LongiControlEnv", "DeterministicTrack", "StochasticTrack"]
+__all__ = [
+    "LongiControlEnv",
+    "DeterministicTrack",
+    "StochasticTrack",
+    "MODeterministicTrack",
+    "MOStochasticTrack",
+]